Broke Combinefall.swift into multiple files
