--- Sources/Combinefall/Catalogs.swift
@@ -1,2 +1,0 @@
-/// A `Catalog` containing up to 20 full English card names.
-public typealias AutocompleteCatalog = Catalog<String>
