--- Sources/Combinefall/Lists.swift
@@ -1,1 +1,0 @@
-public typealias CardList = List<Card>
